refactor: move fineli workflow
