+++ src/call_fineli_api.py
@@ -1,5 +1,4 @@
 import requests
-from src.proccess_fineli_data import get_list_of_nutrition_values
 
 #Fineli and its API are created and maintained by the Finnish Institute for Health and Welfare (THL).
 #more at: https://fineli.fi/fineli/en/index
@@ -23,13 +22,3 @@ def get_data_from_fineli(ingredient, language="en"):
     response = requests.get(url=FINELI_API_URL, params=query_params, headers=headers)
     response = response.json()
     return response
-
-def run_fineli_workflow(query):
-    '''Query Fineli API and clean up the response for further processing.
-
-    Keyword arguments:
-    query -- the name of the ingredient
-    '''
-    fineli_response = get_data_from_fineli(query)
-    clean_fineli_response = get_list_of_nutrition_values(fineli_response)
-    return clean_fineli_response
